desktop/gui/requirements.txt
# UPass GTK GUI Requirements

# Cryptographic libraries
PyNaCl>=1.5.0

# HTTP client for API communication
requests>=2.28.0

# GTK3 Python bindings (system package on most distributions)
# Install via system package manager:
# Ubuntu/Debian: sudo apt install python3-gi python3-gi-cairo gir1.2-gtk-3.0
# Fedora: sudo dnf install python3-gobject gtk3-devel
# Arch: sudo pacman -S python-gobject gtk3

# Optional: For better clipboard support
# pyclip>=0.7.0